\name{ofm-internal}
\title{Service functions and some undocumented functions for the ofm (On Farm Management (OFM) and Experimentation Tools) library}
\alias{.First.lib}
\alias{.onLoad}
\alias{.onAttach}
\alias{.p.val.2tailed}
\alias{coKrig}
\alias{.coKrig_inner_fnc}
\alias{._Global_fnc}
\alias{._Local_fnc}
\alias{._Local_fnc_processed}
\alias{._Local_fnc_fixed}
\alias{._local_fnc_bign}
\alias{._local_fnc_bign_processed}
\alias{._local_fnc_bign_fixed}
\alias{.try_processed_TR1}
\alias{.try_processed_TR2}
\alias{.try_processed_TR3}
\alias{.wrapper_cov}
\alias{.wrapper_check_bign}
\alias{.TR_1}
\alias{.TR_2}
\alias{.TR_3}
\alias{.fnc.time_}
\alias{.wrapper.spplot}
\alias{print.pa}
\description{Internal ofm functions}
\keyword{internal}

