\name{ofm-package}
\alias{ofm-package}
\alias{ofm}
\docType{package}
\title{
On Farm Management (OFM) and Experimentation Tools
}
\description{
 ..........
}
\details{
\tabular{ll}{
Package: \tab ofm\cr
Type: \tab Package\cr
}
Main functions used: \cr
\code{> \link{run}} \cr
\code{> \link{plot.pa}} \cr
Data descriptions: \cr

}
\author{
K.S. Bakar & ... \cr
Maintainer: K.S. Bakar <shuvo.bakar@csiro.com>
}
\references{
.......... \cr
..........
}
\keyword{ package }
\seealso{
Packages 'spacetime', ...; website: \code{http://www.r-project.org/}.
}